refactor: move remaining hooks to standalone files
